refactor(theme): move `getSiteSettings` action to settings store
